refactor: types.go
